refactor(SUP-87): dependencies cleanup part 2 (#7)
* refactor: remove uuid lib

* refactor: move tests folder out of src

* refactor: update jest types
